Refactor mng_schedule cli.py into a cli/ package with one file per command
Split the monolithic cli.py (559 lines) into separate modules:
- cli/group.py: schedule group definition, shared option helpers
- cli/options.py: CLI option classes
- cli/add.py: schedule add command
- cli/remove.py: schedule remove command
- cli/update.py: schedule update command
- cli/list.py: schedule list command + output helpers
- cli/run.py: schedule run command
- cli/commands.py: imports all commands, registers help metadata
