+++ libs/mng_schedule/imbue/mng_schedule/cli/options.py
@@ -1,0 +1,47 @@
+from imbue.mng.cli.common_opts import CommonCliOptions
+
+
+class ScheduleUpdateCliOptions(CommonCliOptions):
+    """Shared options for the schedule add and update subcommands."""
+
+    positional_name: str | None
+    name: str | None
+    command: str | None
+    args: str | None
+    schedule_cron: str | None
+    provider: str | None
+    enabled: bool | None
+    verify: str
+    git_image_hash: str | None
+
+
+class ScheduleAddCliOptions(ScheduleUpdateCliOptions):
+    """Options for the schedule add subcommand.
+
+    These are exactly the same as update--the only difference is whether we error if the name already exists.
+    Name is optional here (unlike update) because a random name can be generated.
+    """
+
+    name: str | None
+    update: bool
+
+
+class ScheduleRemoveCliOptions(CommonCliOptions):
+    """Options for the schedule remove subcommand."""
+
+    names: tuple[str, ...]
+    force: bool
+
+
+class ScheduleListCliOptions(CommonCliOptions):
+    """Options for the schedule list subcommand."""
+
+    all_schedules: bool
+    provider: str
+
+
+class ScheduleRunCliOptions(CommonCliOptions):
+    """Options for the schedule run subcommand."""
+
+    name: str
+    local: bool
